Added GameMode and LobbyType reference tables.

+++ Db/Scripts/003 - Create heroes table.sql
@@ -1,5 +1,5 @@
 ﻿create table Raw.Hero(
-    Id bigint not null,
+    Id int not null,
     Name nvarchar(255) not null,
     LocalizedName nvarchar(255) not null,
     PrimaryAttr varchar(10) not null,
